+++ NEWS.md
@@ -34,6 +34,8 @@
 
 ## Features and bugfixes
 
+* `cnd_footer()` now returns the `footer` field by default, if any.
+
 * `warn()` and `inform()` now signal conditions of classes
   `"rlang_warning"` and `"rlang_message"` respectively.